move _create_study to test_timeline.py from utils.py.
